+++ src/room.cpp
@@ -8,15 +8,15 @@ void Room::create()
     {
         for (int j = 0; j < size.x; j++)
         {
-            int pos_x = pos.x + j;
-            int pos_y = pos.y + i;
+            int pos_x = (int)pos.x + j;
+            int pos_y = (int)pos.y + i;
 
             if (pos_x > level.pos.x && pos_x < level.pos.x + level.size.x && pos_y > level.pos.y && pos_y < level.pos.y + level.size.y)
             {
                 if (i == 0 || i == size.y - 1 || j == 0 || j == size.x - 1)
-                    level.tiles[level.pos.y + pos_y][level.pos.x + pos_x].symbol = wall_symb;
+                    level.tiles[(int)level.pos.y + (int)pos_y][(int)level.pos.x + (int)pos_x].symbol = wall_symb;
                 else
-                    level.tiles[level.pos.y + pos_y][level.pos.x + pos_x].symbol = empty_symb;
+                    level.tiles[(int)level.pos.y + (int)pos_y][(int)level.pos.x + (int)pos_x].symbol = empty_symb;
             }
         }
     }
